+++ SFSimulator.Core/Quests/Quest.cs
@@ -2,15 +2,15 @@
 {
     public class Quest
     {
-        public double Gold { get; set; }
-        public double Experience { get; set; }
-        public double Time { get; set; }
+        public decimal Gold { get; set; }
+        public decimal Experience { get; set; }
+        public decimal Time { get; set; }
         public QuestValue MinimumValues { get; private set; }
         public Item? Item { get; set; }
 
 
 
-        public Quest(QuestValue minimumQuestValue, double gold, double experience, double time, Item? item = null)
+        public Quest(QuestValue minimumQuestValue, decimal gold, decimal experience, decimal time, Item? item = null)
         {
             MinimumValues = minimumQuestValue ?? throw new ArgumentNullException(nameof(minimumQuestValue));
             Gold = gold;
